Address mobile diff PR feedback
