+++ src/vs/sessions/browser/parts/mobile/contributions/media/mobileMultiDiffView.css
@@ -74,6 +74,7 @@
 	padding: 2px;
 	border-radius: 3px;
 	cursor: pointer;
+	touch-action: manipulation;
 }
 
 .mobile-multi-diff-file-chevron::before {
@@ -164,7 +165,7 @@
 /* In the multi-diff view, hunk headers are sticky within each file section. */
 .mobile-multi-diff-file-section .mobile-diff-hunk-header {
 	position: sticky;
-	top: 0;
+	top: 34px;
 	z-index: 1;
 	background: var(--vscode-editor-background, #1e1e1e);
 }
